separate some methods from CommonUtils to io.BufferUtils
